requirements.txt
google-cloud-storage==2.7.0
google-cloud-vision==3.1.0
google-cloud-speech==2.16.0
google-cloud-pubsub==2.13.0
python-magic-bin==0.4.24
python-magic==0.4.27
requests==2.28.2